requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
waitress==2.1.2
Flask-Limiter==3.5.0
pytest==8.0.0
pytest-flask==1.3.0
psycopg2-binary==2.9.9


